requirements.txt
streamlit
pandas
numpy
plotly
openpyxl
pywhatkit
PyPDF2
python-docx
feedparser
requests
streamlit-lottie
supabase==2.16.0
python-dotenv==1.1.1
httpx==0.28.1
pyautogui
twilio
st-supabase-connection==2.1.0